version/docker_health_insurance/requirements.txt
django
djangorestframework
pyjwt
mysqlclient
pymysql